requirements.txt
Django==5.0.4
mysqlclient==2.2.4
channels==4.1.0
channels-redis==4.2.0
kafka-python==2.0.2
daphne==4.1.2
pandas==2.2.1
pyarrow==14.0.1
scikit-learn==1.2.2
xgboost==2.0.3